Job orders pages fixed

+++ database/migrations/2019_11_21_162134_create_cost_approvals_table.php
@@ -13,15 +13,15 @@ class CreateCostApprovalsTable extends Migration
      */
     public function up()
     {
+        // $increments does integer()
         Schema::create('cost_approvals', function (Blueprint $table) {
             $table->increments('id');
-            $table->unsignedBigInteger('requested_by_id')->nullable();
+            $table->integer('requested_by_id')->nullable();
             $table->foreign('requested_by_id')->references('id')->on('employees')->onDelete('SET NULL');
-            $table->unsignedBigInteger('cost_center_id')->nullable();
+            $table->integer('cost_center_id')->nullable();
             $table->foreign('cost_center_id')->references('id')->on('departments')->onDelete('SET NULL');
             $table->string('project_location')->nullable();
             $table->dateTime('date')->nullable();
-
             $table->timestamps();
         });
     }
